adjust check for when to rewrite vc
since type premises also appear when there is no type quantification and this can lead to simplifcations, we also need to rewrite then

currently, we just use an approximation: type quantification occurs or else there is an implication within a quantifier (which can lead to simplifcations). In the future, we may want to make this check more precise, by effectively checking whether the simplified versus non-simplified versions differ.